first calc version
