+++ AMamontov/test/main_tests/CMakeLists.txt
@@ -1,0 +1,15 @@
+file(GLOB SRCS *.cpp)
+
+ADD_EXECUTABLE(main_tests ${SRCS})
+
+TARGET_LINK_LIBRARIES(main_tests 
+	libcalculator     
+	libgtest
+    libgmock)
+
+include_directories (
+  ${BASE_DIR}/include
+)
+
+add_test(NAME main_tests
+         COMMAND main_tests)
