+++ include/string_calc.hpp
@@ -2,7 +2,6 @@
 #define STRING_CALC_HPP
 
 #include <string>
-#include <set>
 
 class StringCalc
 {
@@ -12,19 +11,6 @@ public:
 	~StringCalc();
 	
 	int Add(std::string numbers);
-         
-private:
-        void dropState();
-        void getEndOfTheNextNumber();  
-        void parseSeparators(); 
-        void checkWrongArguments();
-        void setNumbers(std::string numbers); 
-
-        std::set<std::string> separators;
-        std::string m_Numbers; 
-        size_t m_EndPos;
-        size_t m_CurPos;
-        size_t m_Offs;
 };
 
 #endif
